1.1.0

+++ update-nextcloud.sh
@@ -1,26 +1,41 @@
 #!/bin/bash
-# Version 1.0.0
-# Written by Benjamin Wagner, see https://github.com/wagnbeu0/update-nextcloud for updates
+# put the following configuration variables to /etc/nc-updater.conf
+# NC_URL
+# NC_ROOT
+# NC_PUPLIC_URL
+# because in next version of the script it will not contain any user variables
 
-
-# Please set here the URL how your nextcloud is available without HTTPS or HTTP. It is need later when configuring Fulltext Search
-# I recommend to use under /var/www as document root the domain name, like /var/www/intranet.localhost, not the html default folder /var/www/html
-# export NC_URL=html
-export NC_URL=intranet.localhost
+# Please set here the URL how your nextcloud is available without HTTPS or HTTP
+export NC_URL=intranet.mydomain.com
 
 # Please set here the path of your local nextcloud root folder
 export NC_ROOT=/var/www/$NC_URL/nextcloud
 
-echo Update local Apps
+# Please set here the NextCloud Puplic URL:
+export NC_PUPLIC_URL=https://$NC_URL/nextcloud
+
+
+if [[ -f "/etc/nc-updater.conf" ]]; then
+        echo User specific file found, read variables
+        source /etc/nc-updater.conf
+else
+        echo No Config file found, will create /etc/nc-updater.conf now:
+        cat "$0" | head -n 15 | tail -8 >>/etc/nc-updater.conf
+        echo Please check all variables
+fi
+
+echo Update all local Apps
 sudo -u www-data php $NC_ROOT/occ app:update --all
 
 echo Update Nextcloud Main Version
-echo Set Current Main version
+# Set Current Main version
 export current_version=`cat $NC_ROOT/config/config.php | grep -w "version" | cut -c 17,18`
 cd $NC_ROOT
-echo Check available Main version
+
+# Check available Main version
 echo n | sudo -u www-data php $NC_ROOT/updater/updater.phar -vvv | grep channel >/tmp/update-nc.txt
 export available_version=`cat /tmp/update-nc.txt  | cut -c 21,22`
+
 # if no update is available, no version will be displayed, so we will set the installed version as available
 if [[ -z "$available_version" ]]; then
         export available_version=$current_version
@@ -37,15 +52,36 @@ else
 fi
 
 echo Restart Services
-/etc/init.d/elasticsearch stop
-/etc/init.d/elasticsearch start
-/usr/share/elasticsearch/bin/elasticsearch-plugin remove ingest-attachment -s
-/usr/share/elasticsearch/bin/elasticsearch-plugin install ingest-attachment -s --batch
+echo Check if Elasticsearch is installed...
+if [[ -f "/etc/init.d/elasticsearch" ]]; then
+        echo ... Elasticsearch is installed. Restart service
+        systemctl stop elasticsearch
+        systemctl start elasticsearch
+        /usr/share/elasticsearch/bin/elasticsearch-plugin remove ingest-attachment -s
+        /usr/share/elasticsearch/bin/elasticsearch-plugin install ingest-attachment -s --batch
+else
+        echo ... Elasticsearch is not installed
+fi
 
-cd $NC_ROOT
-echo CURL status
-curl https://$NC_URL/nextcloud/apps/richdocumentscode/proxy.php?status
-echo set puplic_wopi_url
-sudo -u www-data php --define apc.enable_cli=1 occ config:app:set richdocuments public_wopi_url --value https://$NC_URL/nextcloud/apps/richdocumentscode/proxy.php?req= &
-echo set wopi_url
-sudo -u www-data php --define apc.enable_cli=1 occ config:app:set richdocuments wopi_url --value https://$NC_URL/nextcloud/apps/richdocumentscode/proxy.php?req= &
+echo Check if RichDocumentsCode and Richdocuments are installed
+if [[ -d "$NC_ROOT/apps/richdocuments" ]]; then
+        echo ... RichDocuments is installed
+
+        if [[ -d "$NC_ROOT/apps/richdocumentscode" ]]; then
+                echo ... RichDocumentsCode is installed
+
+                cd $NC_ROOT
+                echo CURL status
+                curl $NC_PUPLIC_URL/apps/richdocumentscode/proxy.php?status
+                echo set puplic_wopi_url
+                sudo -u www-data php --define apc.enable_cli=1 occ config:app:set richdocuments public_wopi_url --value $NC_PUPLIC_URL/apps/richdocumentscode/proxy.php?req= &
+                echo set wopi_url
+                sudo -u www-data php --define apc.enable_cli=1 occ config:app:set richdocuments wopi_url --value $NC_PUPLIC_URL/apps/richdocumentscode/proxy.php?req= &
+                echo Active Config
+                sudo -u www-data php --define apc.enable_cli=1 occ richdocuments:activate-config &
+        else
+                echo ... RichDocumentsCode is not installed
+        fi
+else
+        echo ... RichDocuments is not installed
+fi
